Fix FSE encoding/decoding: pre-allocate buffer, sync memcpy, correct state bit width
Key fixes in cuda_zstd_fse.cu:
1. Pre-allocate encoder output buffer before bit writes
2. Use synchronous cudaMemcpy for local buffer (prevent use-after-free)
3. Write final state with table_log+1 bits (states are in [table_size, 2*table_size))
4. Decoder reads table_log+1 bits and converts from encoder range to table index

All 8 FSE comprehensive tests now pass.

+++ tests/test_data_integrity_comprehensive.cu
@@ -158,8 +158,8 @@ bool test_backtracking_integrity(const char *pattern_name,
       }
       // Coverage may not match exactly due to segment-based storage
     } else if (total_coverage != input_size) {
-      printf("    FAIL: Coverage mismatch! Expected %zu, got %llu\n",
-             input_size, total_coverage);
+      printf("    FAIL: Coverage mismatch! Expected %zu, got %lu\n", input_size,
+             (unsigned long)total_coverage);
       passed = false;
     }
   }
